rag/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
qdrant-client>=1.7.0
google-genai>=0.2.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
redis>=5.0.0
